python/requirements.txt
tensorflow==2.20.0
numpy
opencv-python
pillow
scikit-image
tqdm
matplotlib
